@@ -54,7 +54,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
     "@types/jest": "^29.5.0",
-    "@types/react": "^16.9.37",
+    "@types/react": "16.9.41",
     "@types/react-dom": "^16.9.8",
     "@types/react-virtualized": "9.21.10",
     "eslint": "^8.47.0",
@@ -63,13 +63,13 @@
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
-    "joi": "^17.9.1",
+    "joi": "^17.12.1",
     "lerna": "^6.6.1",
     "lint-staged": "^10.4.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
-    "react": "^16.13.1",
-    "react-dom": "16.13.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "rxjs": "^6.6.7",
     "sort-package-json": "^1.50.0",
     "ts-jest": "^29.1.0",
